todoApp/requirements.txt
absl-py==1.4.0
annotated-types==0.5.0
anyio==3.7.1
attrs==23.1.0
bcrypt==4.0.1
cffi==1.15.1
click==8.1.6
colorama==0.4.6
contourpy==1.1.0
cryptography==41.0.2
cvzone==1.5.6
cycler==0.11.0
ecdsa==0.18.0
exceptiongroup==1.1.2
fastapi==0.100.1
flatbuffers==23.5.26
fonttools==4.41.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-resources==6.0.0
kiwisolver==1.4.4
matplotlib==3.7.2
mediapipe==0.10.2
packaging==23.1
passlib==1.7.4
protobuf==3.20.3
pyasn1==0.5.0
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
SQLAlchemy==2.0.19
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.1
zipp==3.16.2
